Define abstract Bicycle class
Define abstract Bicycle class and derived class BicycleWhipple. This
abstract class defines a number of member functions and variables that
are shared between BicycleWhipple and the (not yet implemented) class
BicycleKinematic. The implementation of the previous Bicycle class has
been split between Bicycle and BicycleWhipple.

Other changes include:
 - Removal of integrate_state(state_t, input_t, real_t) prototype
 - Rename and prototype change of update_auxiliary_state(state_t,
   auxiliary_state_t)
 - Removal of private function initialize_state_space_matrices()
 - Refactor of state space matrix computation
 - Removal of no input integration stepper
